Create Floyd's-Hare-And-Turtle.c
Floyd's Hare And Turtle is a cycle detection algorithm which looks for cycles in an array or linked list by creating two pointers and traversing them throughout the array and looking for similar values at different. This leads to an algorithm capable of finding cycles in an array with a Big O(n).